\name{crossR-package}
\alias{crossR-package}
\alias{crossR}
\docType{package}
\title{
  crossR: Simulate agent-based models for animal-road systems
}
\description{
  Builds simulation of animal movement between seasonal ranges with road interactions. Includes sample data using publicly-available Western Arctic Herd seasonal ranges. 
}
\details{
To see a detailed vignette:
```
vignette("crossR")
```
}
\author{
Maintainer: Kayleigh Chalkowski <Kayleigh.Chalkowski@tamuk.edu>
}
\references{
Joly K. 2026. Western Arctic Herd caribou annual and seasonal range distributions. National Park Service
https://irma.nps.gov/DataStore/Reference/Profile/2318272
}
\keyword{ crossR }
\seealso{
}
\examples{
  \dontrun{
  }
}
