\name{Visualize_Tracks}
\alias{Visualize_Tracks}
\title{Animate Simulated Tracks}
\description{
This function takes tracking output from Range_Simulate (using out.opts option "tracking") and animates movement of each simulated individual. Color change represents shift in seasonal movement strategy.     

Animation is saved to input path.
}
\usage{

}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{tracking}{tracking item from Range_Simulate list output}
  \item{path}{String, path with filename for output animation}
  \item{spatial_input}{List, output from Format_Spatial_Data}
  \item{country}{String, country name matching string format within rnaturalearth}
  \item{plot_buffer}{Numeric vector, additional buffer around bbox to improve aesthetic}
}
\value{

}
\references{
}
\author{
Kayleigh Chalkowski, Kayleigh.Chalkowski@tamuk.edu
}
\note{
}

\seealso{
}
\examples{

}
