\name{Vizualize_Tracks}
\alias{Vizualize_Tracks}
\title{
Vizualize_Tracks}
\description{

}
\usage{

}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{range_list}{a list}
  \item{rdcoords}{a matrix of xy coordinates}
  \item{inc}{an integer, the spatial resolution for the simulation}
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
\item{grid_list}{list of grid objects}
\item{other outputs}{other stuff}
}
\references{
}
\author{
Kayleigh Chalkowski
}
\note{
}

\seealso{
}
\examples{

}
