\name{Range_Simulate}
\alias{Range_Simulate}
\title{
Run simulation of animal movement and road interaction
}
\description{

}
\usage{
Range_Simulate(spatial_input,
							 rdcoords,
							 mv_jday,
							 jday_max,
							 N0,
							 inc,
							 dist_start,
							 out.opts=c("tracking")
							 )
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{range_list}{a list}
  \item{rdcoords}{a matrix of xy coordinates}
  \item{inc}{an integer, the spatial resolution for the simulation}
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
\item{grid_list}{list of grid objects}
\item{other outputs}{other stuff}
}
\references{
}
\author{
Kayleigh Chalkowski
}
\note{
}

\seealso{
}
\examples{

}
